Allow for setting multiple LEDs to different colors at the same time
The design for this essentially throws out the blocking feature from
earlier, but I'll leave it in the library in case it becomes useful down
the line. The new way of doing things is explicit waits have to be
inserted into the JSON. I'm thinking I will make a GUI editor for this
file to make that a little more intuitive. With what is in the current
example JSON I think it's still pretty easy to follow, but it's easy to
imagine a case where the LEDs are not necessarily changing in sync with
each other, in which case figuring out what the waits should be could be
complicated.